JoinKeysignViewModel refactored
